requirements.txt
# VentureLens AI - Requirements
# Python 3.12+

# Core Django
Django==5.1.4
django-environ==0.11.2
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.10

# HTMX integration
django-htmx==1.21.0

# AI / LLM
openai==1.58.1          # Compatible with Groq API
groq==0.13.1            # Direct Groq SDK

# Data & Trends
pytrends==4.9.2
requests==2.32.3

# Forms & Validation
django-crispy-forms==2.3
crispy-tailwind==1.0.3

# Utilities
python-dotenv==1.0.1
Pillow==11.0.0
whitenoise==6.8.2        # Static files in production

# Cache (optional Redis)
django-redis==5.4.0

# Dev tools
django-debug-toolbar==4.4.6

# Production
gunicorn==23.0.0

# Security
django-cors-headers==4.6.0
